Move S3Verifier inner classes into their own package
