+++ src/main/java/com/flightstats/hub/dao/aws/S3Verifier.java
@@ -1,20 +1,16 @@
 package com.flightstats.hub.dao.aws;
 
-import com.flightstats.hub.app.HubProperties;
 import com.flightstats.hub.app.HubServices;
 import com.flightstats.hub.cluster.*;
 import com.flightstats.hub.dao.ChannelService;
-import com.flightstats.hub.dao.ContentDao;
-import com.flightstats.hub.dao.QueryResult;
+import com.flightstats.hub.dao.aws.s3Verifier.MissingContentFinder;
 import com.flightstats.hub.dao.aws.s3Verifier.VerifierMetrics;
-import com.flightstats.hub.metrics.ActiveTraces;
+import com.flightstats.hub.dao.aws.s3Verifier.VerifierConfig;
+import com.flightstats.hub.dao.aws.s3Verifier.VerifierRange;
+import com.flightstats.hub.dao.aws.s3Verifier.VerifierRangeLookup;
 import com.flightstats.hub.metrics.MetricsService;
-import com.flightstats.hub.metrics.Traces;
 import com.flightstats.hub.model.*;
-import com.flightstats.hub.spoke.SpokeStore;
-import com.flightstats.hub.spoke.SpokeStoreConfig;
 import com.flightstats.hub.util.HubUtils;
-import com.flightstats.hub.util.RuntimeInterruptedException;
 import com.flightstats.hub.util.Sleeper;
 import com.flightstats.hub.util.TimeUtil;
 import com.google.common.annotations.VisibleForTesting;
@@ -26,23 +22,20 @@ import com.sun.jersey.api.client.Client;
 import com.sun.jersey.api.client.ClientResponse;
 import org.apache.curator.framework.CuratorFramework;
 import org.joda.time.DateTime;
-import org.joda.time.Duration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.SortedSet;
-import java.util.TreeSet;
-import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
 
 @Singleton
 public class S3Verifier {
-    static final String LAST_SINGLE_VERIFIED = "/S3VerifierSingleLastVerified/";
+    public static final String LAST_SINGLE_VERIFIED = "/S3VerifierSingleLastVerified/";
     private final static Logger logger = LoggerFactory.getLogger(S3Verifier.class);
     private static final String LEADER_PATH = "/S3VerifierSingleService";
 
-    private final S3VerifierConfig s3VerifierConfig;
+    private final VerifierConfig verifierConfig;
 
     private final ExecutorService channelThreadPool;
     private final LastContentPath lastContentPath;
@@ -53,7 +46,7 @@ public class S3Verifier {
     private final CuratorFramework curator;
     private final MetricsService metricsService;
     private final MissingContentFinder missingContentFinder;
-    private final S3VerifierRangeLookup verifierRangeLookup;
+    private final VerifierRangeLookup verifierRangeLookup;
 
     @Inject
     public S3Verifier(LastContentPath lastContentPath,
@@ -64,8 +57,8 @@ public class S3Verifier {
                       CuratorFramework curator,
                       MetricsService metricsService,
                       MissingContentFinder missingContentFinder,
-                      S3VerifierRangeLookup verifierRangeLookup,
-                      S3VerifierConfig s3VerifierConfig,
+                      VerifierRangeLookup verifierRangeLookup,
+                      VerifierConfig verifierConfig,
                       @Named("s3VerifierChannelThreadPool") ExecutorService channelThreadPool) {
         this.lastContentPath = lastContentPath;
         this.channelService = channelService;
@@ -74,12 +67,12 @@ public class S3Verifier {
         this.zooKeeperState = zooKeeperState;
         this.curator = curator;
         this.metricsService = metricsService;
-        this.s3VerifierConfig = s3VerifierConfig;
+        this.verifierConfig = verifierConfig;
         this.channelThreadPool = channelThreadPool;
         this.missingContentFinder = missingContentFinder;
         this.verifierRangeLookup = verifierRangeLookup;
 
-        if (s3VerifierConfig.isEnabled()) {
+        if (verifierConfig.isEnabled()) {
             HubServices.register(new S3ScheduledVerifierService(), HubServices.TYPE.AFTER_HEALTHY_START, HubServices.TYPE.PRE_STOP);
         }
     }
@@ -93,7 +86,7 @@ public class S3Verifier {
                     channelThreadPool.submit(() -> {
                         String name = Thread.currentThread().getName();
                         Thread.currentThread().setName(name + "|" + channel.getDisplayName());
-                        String url = s3VerifierConfig.getEndpointUrlGenerator().apply(channel.getDisplayName());
+                        String url = verifierConfig.getEndpointUrlGenerator().apply(channel.getDisplayName());
                         logger.debug("calling {}", url);
                         ClientResponse post = null;
                         try {
@@ -153,123 +146,11 @@ public class S3Verifier {
         }
     }
 
-    @VisibleForTesting
-    static class S3VerifierRangeLookup {
-        private final S3VerifierConfig s3VerifierConfig;
-        private final SpokeStoreConfig spokeWriteStoreConfig;
-
-        private final LastContentPath lastContentPath;
-        private final ChannelService channelService;
-
-        @Inject
-        public S3VerifierRangeLookup(LastContentPath lastContentPath,
-                          ChannelService channelService,
-                          S3VerifierConfig s3VerifierConfig,
-                          @Named("spokeWriteStoreConfig") SpokeStoreConfig spokeWriteStoreConfig) {
-            this.lastContentPath = lastContentPath;
-            this.channelService = channelService;
-            this.s3VerifierConfig = s3VerifierConfig;
-            this.spokeWriteStoreConfig = spokeWriteStoreConfig;
-        }
-
-        VerifierRange getSingleVerifierRange(DateTime now, ChannelConfig channelConfig) {
-            MinutePath spokeTtlTime = getSpokeTtlPath(now);
-            now = channelService.getLastUpdated(channelConfig.getDisplayName(), new MinutePath(now)).getTime();
-            DateTime start = now.minusMinutes(1);
-            MinutePath endPath = new MinutePath(start);
-            MinutePath defaultStart = new MinutePath(start.minusMinutes(s3VerifierConfig.getOffsetMinutes()));
-            MinutePath startPath = (MinutePath) lastContentPath.get(channelConfig.getDisplayName(), defaultStart, LAST_SINGLE_VERIFIED);
-            if (channelConfig.isLive() && startPath.compareTo(spokeTtlTime) < 0) {
-                startPath = spokeTtlTime;
-            }
-            return VerifierRange.builder()
-                    .channelConfig(channelConfig)
-                    .startPath(startPath)
-                    .endPath(endPath)
-                    .build();
-        }
-
-        private MinutePath getSpokeTtlPath(DateTime now) {
-            return new MinutePath(now.minusMinutes(spokeWriteStoreConfig.getTtlMinutes() - 2));
-        }
-    }
-
-    @VisibleForTesting
-    static class MissingContentFinder {
-        private final S3VerifierConfig s3VerifierConfig;
-        private final ExecutorService queryThreadPool;
-        private final ContentDao spokeWriteContentDao;
-        private final ContentDao s3SingleContentDao;
-        private final MetricsService metricsService;
-
-        @Inject
-        public MissingContentFinder(@Named(ContentDao.WRITE_CACHE) ContentDao spokeWriteContentDao,
-                                    @Named(ContentDao.SINGLE_LONG_TERM) ContentDao s3SingleContentDao,
-                                    S3VerifierConfig s3VerifierConfig,
-                                    MetricsService metricsService,
-                                    @Named("s3VerifierQueryThreadPool") ExecutorService queryThreadPool) {
-            this.spokeWriteContentDao = spokeWriteContentDao;
-            this.s3SingleContentDao = s3SingleContentDao;
-            this.metricsService = metricsService;
-            this.s3VerifierConfig = s3VerifierConfig;
-            this.queryThreadPool = queryThreadPool;
-        }
-
-        @VisibleForTesting
-        protected SortedSet<ContentKey> getMissing(MinutePath startPath, MinutePath endPath, String channelName) {
-            long timeout = s3VerifierConfig.getBaseTimeoutMinutes();
-            QueryResult queryResult = new QueryResult(1);
-            SortedSet<ContentKey> s3Keys = new TreeSet<>();
-            SortedSet<ContentKey> spokeKeys = new TreeSet<>();
-            TimeQuery.TimeQueryBuilder builder = TimeQuery.builder()
-                    .channelName(channelName)
-                    .startTime(startPath.getTime())
-                    .unit(TimeUtil.Unit.MINUTES);
-            if (endPath != null) {
-                Duration duration = new Duration(startPath.getTime(), endPath.getTime());
-                timeout += duration.getStandardDays();
-                builder.limitKey(ContentKey.lastKey(endPath.getTime()));
-            }
-            TimeQuery timeQuery = builder.build();
-            try {
-                CountDownLatch latch = new CountDownLatch(2);
-                runInQueryPool(ActiveTraces.getLocal(), latch, () -> spokeKeys.addAll(spokeWriteContentDao.queryByTime(timeQuery)));
-                runInQueryPool(ActiveTraces.getLocal(), latch, () -> s3Keys.addAll(s3SingleContentDao.queryByTime(timeQuery)));
-                latch.await(timeout, TimeUnit.MINUTES);
-                if (latch.getCount() != 0) {
-                    logger.error("s3 verifier timed out while finding missing items, write queue is backing up");
-                    metricsService.increment(VerifierMetrics.TIMEOUT.getName());
-                    return new TreeSet<>();
-                }
-                spokeKeys.removeAll(s3Keys);
-                if (spokeKeys.size() > 0) {
-                    logger.info("missing items {} {}", channelName, queryResult.getContentKeys());
-                }
-                return spokeKeys;
-//            throw new FailedQueryException("unable to query spoke");
-            } catch (InterruptedException e) {
-                throw new RuntimeInterruptedException(e);
-            }
-        }
-
-        private void runInQueryPool(Traces traces, CountDownLatch countDownLatch, Runnable runnable) {
-            queryThreadPool.submit(() -> {
-                ActiveTraces.setLocal(traces);
-                try {
-                    runnable.run();
-                } finally {
-                    countDownLatch.countDown();
-                }
-            });
-        }
-    }
-
-
     private void incrementMetric(VerifierMetrics verifierMetric) {
         metricsService.increment(verifierMetric.getName());
     }
-    private class S3ScheduledVerifierService extends AbstractScheduledService implements Lockable {
 
+    private class S3ScheduledVerifierService extends AbstractScheduledService implements Lockable {
         @Override
         protected void runOneIteration() throws Exception {
             CuratorLock curatorLock = new CuratorLock(curator, zooKeeperState, LEADER_PATH);
@@ -277,7 +158,7 @@ public class S3Verifier {
         }
 
         protected Scheduler scheduler() {
-            return Scheduler.newFixedDelaySchedule(0, s3VerifierConfig.getOffsetMinutes(), TimeUnit.MINUTES);
+            return Scheduler.newFixedDelaySchedule(0, verifierConfig.getOffsetMinutes(), TimeUnit.MINUTES);
         }
 
         @Override
@@ -286,7 +167,7 @@ public class S3Verifier {
             while (leadership.hasLeadership()) {
                 long start = System.currentTimeMillis();
                 verifySingleChannels();
-                long sleep = TimeUnit.MINUTES.toMillis(s3VerifierConfig.getOffsetMinutes()) - (System.currentTimeMillis() - start);
+                long sleep = TimeUnit.MINUTES.toMillis(verifierConfig.getOffsetMinutes()) - (System.currentTimeMillis() - start);
                 logger.debug("sleeping for {} ms", sleep);
                 Sleeper.sleep(Math.max(0, sleep));
                 logger.debug("waking up after sleep");
