+++ src/main/java/com/flightstats/hub/dao/aws/s3verifier/VerifierConfig.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.dao.aws;
+package com.flightstats.hub.dao.aws.s3Verifier;
 
 import lombok.Builder;
 import lombok.Value;
@@ -7,7 +7,7 @@ import java.util.function.Function;
 
 @Value
 @Builder
-public class S3VerifierConfig {
+public class VerifierConfig {
     boolean enabled;
 
     long baseTimeoutMinutes;
