+++ src/main/java/com/flightstats/hub/dao/aws/s3verifier/VerifierRange.java
@@ -1,4 +1,4 @@
-package com.flightstats.hub.dao.aws;
+package com.flightstats.hub.dao.aws.s3Verifier;
 
 import com.flightstats.hub.model.ChannelConfig;
 import com.flightstats.hub.model.MinutePath;
@@ -7,7 +7,7 @@ import lombok.Value;
 
 @Value
 @Builder
-class VerifierRange {
+public class VerifierRange {
     private final MinutePath startPath;
     private final MinutePath endPath;
     private final ChannelConfig channelConfig;
